major part of continuous time rewrite
compiles but does not link

diagonals now are a map. most of the references to the old setup are
gone

make_svd_double DOES NOT WORK YET. It has a bug in the intervals
computed. it needs an additional t2 to keep track of where it has been
computed already;

all of the load/save functions are still the old ones. they compile due
to the conversion int->double and will screw up everything since map[i]
creates the element. sweet.

there's probably a ton of bugs left